# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cinema AI Translator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: sessions
    description: Translation session management
  - name: translate
    description: Audio translation endpoints

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List all translation sessions for the current user
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranslationSession"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new translation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationSession"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a session with its logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationSessionWithLogs"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /sessions/{id}/logs:
    get:
      operationId: listSessionLogs
      tags: [sessions]
      summary: List translation logs for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranslationLog"

  /sessions/{id}/export:
    get:
      operationId: getSessionExport
      tags: [sessions]
      summary: Export session logs as text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionExport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /translate/chunk:
    post:
      operationId: translateChunk
      tags: [translate]
      summary: Translate an audio chunk — STT, translate, TTS, return audio + transcript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChunkInput"
      responses:
        "200":
          description: Translation result with audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /sessions/stats:
    get:
      operationId: getSessionStats
      tags: [sessions]
      summary: Get session stats summary for dashboard
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStats"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    TranslationSession:
      type: object
      required: [id, name, targetLanguage, createdAt, logCount]
      properties:
        id:
          type: integer
        name:
          type: string
        targetLanguage:
          type: string
        createdAt:
          type: string
          format: date-time
        logCount:
          type: integer

    SessionInput:
      type: object
      required: [name, targetLanguage]
      properties:
        name:
          type: string
          minLength: 1
        targetLanguage:
          type: string
          minLength: 1

    TranslationLog:
      type: object
      required: [id, sessionId, originalText, translatedText, sourceLanguage, targetLanguage, timestamp]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string
        timestamp:
          type: string
          format: date-time

    LogInput:
      type: object
      required: [originalText, translatedText, sourceLanguage, targetLanguage]
      properties:
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        targetLanguage:
          type: string

    TranslationSessionWithLogs:
      type: object
      required: [id, name, targetLanguage, createdAt, logCount, logs]
      properties:
        id:
          type: integer
        name:
          type: string
        targetLanguage:
          type: string
        createdAt:
          type: string
          format: date-time
        logCount:
          type: integer
        logs:
          type: array
          items:
            $ref: "#/components/schemas/TranslationLog"

    ChunkInput:
      type: object
      required: [audio, targetLanguage, sessionId]
      properties:
        audio:
          type: string
          description: Base64-encoded audio data (WebM or WAV)
        targetLanguage:
          type: string
        sessionId:
          type: integer

    ChunkResult:
      type: object
      required: [originalText, translatedText, sourceLanguage, audioBase64, latencyMs]
      properties:
        originalText:
          type: string
        translatedText:
          type: string
        sourceLanguage:
          type: string
        audioBase64:
          type: string
          description: Base64-encoded MP3 audio of the translation
        latencyMs:
          type: integer

    SessionExport:
      type: object
      required: [content, filename]
      properties:
        content:
          type: string
        filename:
          type: string

    SessionStats:
      type: object
      required: [totalSessions, totalLogs, recentSessions]
      properties:
        totalSessions:
          type: integer
        totalLogs:
          type: integer
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/TranslationSession"